Deduplicate LanguageUtils classes
