+++ src/main/java/io/codiga/plugins/jetbrains/actions/create_recipe/AssistantCreateRecipeAction.java
@@ -9,7 +9,7 @@ import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.components.JBLabel;
 import io.codiga.api.type.LanguageEnumeration;
-import io.codiga.plugins.jetbrains.graphql.LanguageUtils;
+import io.codiga.plugins.jetbrains.utils.LanguageUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
