+++ src/main/java/io/codiga/plugins/jetbrains/dependencies/DependencyManagement.java
@@ -5,7 +5,7 @@ import com.intellij.openapi.vfs.VirtualFile;
 import io.codiga.api.type.LanguageEnumeration;
 import io.codiga.plugins.jetbrains.model.Dependency;
 import com.google.common.collect.ImmutableList;
-import io.codiga.plugins.jetbrains.graphql.LanguageUtils;
+import io.codiga.plugins.jetbrains.utils.LanguageUtils;
 
 import java.util.List;
